requirements.txt
alembic==1.12.0
blinker==1.6.3
click==8.1.7
Flask==3.0.0
Flask-Migrate==2.5.2
Flask-SQLAlchemy==3.1.1
greenlet==3.0.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.21
typing_extensions==4.8.0
Werkzeug==3.0.0
zipp==3.17.0
